Support for multithreaded drivers (kb and display)

+++ include/kernel/memory/physical.hpp
@@ -27,7 +27,7 @@ extern "C" void tlb_flush();
 // Page Table functions
 page_directory_t pg_directory_setup();
 void kernel_pg_setup();
-void map_vaddr_page(uintptr_t virtual_address);
+void map_vaddr_page(uint32_t virtual_address);
 void map_vaddr_page(uint32_t virtual_address, uint32_t phys_address);
